requirements.txt
django
django-mptt
djangorestframework
djangorestframework-recursive